requirements.txt
PyMuPDF
pandasai
beautifulsoup4
fuzzywuzzy
mysql-connector-python
selenium
webdriver_manager
streamlit-option-menu
streamlit_extras
